PNIP-10 quase quase

+++ src/main/resources/db/migration/V202312180201__create_table_unidade_usuario.sql
@@ -1,9 +1,9 @@
 CREATE TABLE unidade_usuario (
                                  id SERIAL PRIMARY KEY,
-                                 id_unidade INT,
-                                 id_usuario INT,
+                                 id_unidade BIGINT,
+                                 id_usuario BIGINT,
                                  ativo BOOLEAN,
-                                 id_permissao INT,
+                                 id_permissao BIGINT,
                                  FOREIGN KEY (id_unidade) REFERENCES unidade(id),
                                  FOREIGN KEY (id_usuario) REFERENCES usuario(id),
                                  FOREIGN KEY (id_permissao) REFERENCES permissao(id)
